Converted Book component to TS and had both the home page and book page use it
